requirements.txt
wikitools>=1.2
pyyaml>=3.11
